+++ sandbox/chat-notification-pubnub-example/facade/src/models/index.ts
@@ -1,0 +1,3 @@
+export * from './pubnub-message.model';
+export * from './pubnubnotification.model';
+export * from './pubnub-message-recipient.model';