+++ services/search-service/migrations/sqls/20211003143854-init-down.sql
@@ -1,0 +1,2 @@
+drop table main.recent_search;
+drop table main.search_query;